saving before switching branches

+++ src/solver/impls/cvode/cvode.cxx
@@ -45,11 +45,12 @@
 #define ZERO        RCONST(0.)
 #define ONE         RCONST(1.0)
 
-typedef long CVINT;
+typedef int CVINT;
 
 static int cvode_rhs(BoutReal t, N_Vector u, N_Vector du, void *user_data);
 static int cvode_bbd_rhs(CVINT Nlocal, BoutReal t, N_Vector u, N_Vector du, 
 			 void *user_data);
+
 
 static int cvode_pre(BoutReal t, N_Vector yy, N_Vector yp,
 		     N_Vector rvec, N_Vector zvec,
